Add new logic to Observable.java, Observer.java, TriangleObservable.java, TriangleStore.java
