+++ src/main/java/org/training/triangles/Observable.java
@@ -1,4 +1,8 @@
 package org.training.triangles;
 
 public interface Observable {
+
+    void attach(Observer observer);
+    void detach(Observer observer);
+    void notifyObservers();
 }
