+++ src/main/java/org/training/triangles/Observer.java
@@ -1,8 +1,5 @@
 package org.training.triangles;
 
 public interface Observer {
-
-    void attach(Observer observer);
-    void detach(Observer observer);
-    void notifyObservers();
+    void update(TriangleObservable event);
 }
